+++ packages/components/components/elvis-chip/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Chip Changelog
 
+## 1.2.2 (03.03.22)
+
+### Patch
+
+- Now using correct type for webcomponent wrapper for better internal IntelliSense.
+
 ## 1.2.1 (18.02.22)
 
 ### Patch