+++ packages/components/components/elvis-divider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Divider Changelog
 
+## 1.2.4 (03.03.22)
+
+### Patch
+
+- Now using correct type for webcomponent wrapper for better internal IntelliSense.
+
 ## 1.2.3 (28.02.22)
 
 ### Bug fix